scrap controller 내용 수정, courseId long>int 로 필드 변수형 수정

+++ coursemos_spring_boot/src/main/java/com/example/demo/dao/ScrapDao.java
@@ -8,7 +8,7 @@ import com.example.demo.domain.Scrap;
 public interface ScrapDao {
 	List<Scrap> getAllScrap() throws DataAccessException;
 	List<Scrap> getScrapByMemberId(Long memberId) throws DataAccessException;
-	Scrap getScrapByPrimaryKey(Long memberId, Long courseId) throws DataAccessException;
+	Scrap getScrapByPrimaryKey(Long memberId, int courseId) throws DataAccessException;
 	void insertScrap(Scrap scrap) throws DataAccessException;
 	void deleteScrap(Scrap scrap) throws DataAccessException;
 }
